T5 Tokenization Fixed
Training of T5, dataset creation: Instead of first splitting by spaCy and mask a word in the text sequence, the text sequence is first encoded, and the word boundaries is the intersection between T5 tokenization boundaries and spaCy tokenization boundaries. A random "word" in that sense is then masked to create a masked sequence.