Pred & Gold fixed
Previous computation of perplexity-related scores had a bug with log-softmax. It was also confusing since the logits outputted by model call and the scores outputted by model.generate are in different format.   The two outputs are now first reformated into the same format, and then the same method is used to compute the scores.